+++ src/js/const/will-request-states.const.js
@@ -1,0 +1,8 @@
+export const WILL_REQUEST_STATES = Object.freeze({
+  deleted: -3,
+  rejected: -1,
+  submitted: 0,
+  approved: 1,
+  notified: 2,
+  released: 3,
+})