+++ src/main.js
@@ -9,12 +9,15 @@ import { store } from '@/vuex'
 import { vueRoutes } from '@/vue-router/routes'
 import { CONFIG } from '@/config'
 import { i18n } from '@/i18n'
-import { ripple } from '@/vue/directives/ripple'
+import { ripple, clickOutside } from '@/vue/directives'
 import { createApp, h, getCurrentInstance } from 'vue'
-import { useFormatDate, useGlobalizeUserRole } from '@/vue/composables'
+import {
+  useFormatDate,
+  useGlobalizeUserRole,
+  useGlobalizeWillRequestStatus,
+} from '@/vue/composables'
 import { ErrorHandler } from '@/js/helpers/error-handler'
 
-import VueClickAway from 'vue3-click-away'
 import Maska from 'maska'
 
 const app = createApp({
@@ -30,6 +33,7 @@ const app = createApp({
     } = useFormatDate()
 
     const { globalizeUserRole } = useGlobalizeUserRole()
+    const { globalizeWillRequestStatus } = useGlobalizeWillRequestStatus()
 
     /* eslint-disable max-len */
     app.appContext.config.globalProperties.$fd = formatDate
@@ -38,6 +42,7 @@ const app = createApp({
     app.appContext.config.globalProperties.$fcalend = formatCalendar
     app.appContext.config.globalProperties.$fcalendi = formatCalendarInline
     app.appContext.config.globalProperties.$globalizeUserRole = globalizeUserRole
+    app.appContext.config.globalProperties.$globalizeWillRequestStatus = globalizeWillRequestStatus
     /* eslint-enable max-len */
   },
   render: () => h(App),
@@ -47,13 +52,13 @@ app
   .use(store)
   .use(router)
   .use(i18n)
-  .use(VueClickAway)
   .use(Maska)
 
 app.config.globalProperties.$routes = vueRoutes
 app.config.globalProperties.$config = CONFIG
 
 app.directive('ripple', ripple)
+app.directive('click-outside', clickOutside)
 
 app.component('AppButton', AppButton)