+++ src/vue/common/NoDataMessage.vue
@@ -3,6 +3,10 @@
     class="no-data-message"
     :class="{ 'no-data-message--center': center }"
   >
+    <i
+      :class="`mdi mdi-${iconName}`"
+      class="no-data-message__icon"
+    />
     <h2 class="no-data-message__title">
       {{ title }}
     </h2>
@@ -18,6 +22,8 @@ export default {
   name: 'no-data-message',
 
   props: {
+    iconName: { type: String, default: 'trending-up' },
+
     title: { type: String, required: true },
 
     message: { type: String, required: true },
@@ -41,17 +47,13 @@ export default {
   }
 }
 
-.no-data-message__title {
-  margin-bottom: 0.8rem;
-  font-size: 1.4rem;
-}
-
-.no-data-message__message {
-  font-size: 1.2rem;
-}
+.no-data-message__title { margin-bottom: 0.8rem; }
 
 .no-data-message__title,
-.no-data-message__message {
-  color: $col-no-data-message-text;
+.no-data-message__message { color: $col-no-data-message-text; }
+
+.no-data-message__icon {
+  color: $col-no-data-message-icon-color;
+  font-size: 6.4rem;
 }
 </style>