+++ src/vue/pages/WillRequests/WillRequestsDetails.vue
@@ -1,11 +1,11 @@
 <template>
-  <div class="will-requests-list">
+  <div class="will-requests-details">
     <!--  -->
   </div>
 </template>
 
 <script>
 export default {
-  name: 'will-requests-list',
+  name: 'will-requests-details',
 }
 </script>